requirements.txt
flask
faiss-cpu
cohere
transformers
sentence-transformers
pdfminer.six
langdetect
